+++ docs/pages/world/modules/_meta.js
@@ -1,4 +1,5 @@
 export default {
   "keyswithvalue": "Keys with Value",
+  "keysintable": "Keys in Table",
   "erc721": "ERC-721 (NFT)"
 };