+++ packages/explorer/src/components/LatestBlock.tsx
@@ -1,25 +1,26 @@
 import { useBlockNumber } from "wagmi";
+import { Skeleton } from "./ui/Skeleton";
 
 export function LatestBlock() {
   const { data: block } = useBlockNumber({
     watch: true,
   });
 
-  if (block === undefined || block === 0n) {
-    return;
-  }
-
   return (
-    <div className="inline-block">
-      <div className="flex items-center justify-between text-xs font-extrabold text-green-600">
-        <span
-          className="mr-2 inline-block h-[8px] w-[8px] animate-pulse rounded-full"
-          style={{
-            background: "rgb(64, 182, 107)",
-          }}
-        ></span>
-        <span className="opacity-70">{block.toString()}</span>
-      </div>
+    <div className="inline-block w-[50px]">
+      {block ? (
+        <div className="flex items-center justify-end text-xs font-extrabold text-green-600">
+          <span
+            className="mr-2 inline-block h-[8px] w-[8px] animate-pulse rounded-full"
+            style={{
+              background: "rgb(64, 182, 107)",
+            }}
+          ></span>
+          <span className="opacity-70">{block.toString()}</span>
+        </div>
+      ) : (
+        <Skeleton className="h-[10px]" />
+      )}
     </div>
   );
 }